Fix getStroke when there are multiple duplicates
Using naive approaches to adding points, quite often we get hundreds of duplicates. Occasionally if duplicates happen in the correct sequence near the beginning of the points array, `i` would outpace the number of elements in strokePoints.

I'm unsure if there are better solutions to this but this change makes sense within the scope of the loop and fixes the problems in my project.